test(sum): split `SumFixture` into separate partial classes for cache and list sources
